Version: 0.4.4:bugs fixed

+++ geoutils/modisAPI.py
@@ -38,7 +38,7 @@ def main():
     parser.add_argument("--path_aoi", help="Path for AOI geojson. ", type=str)
     parser.add_argument("--crs_aoi", help="CRS for AOI geojson. Default: 4326 ", type=int, default=4326)
     parser.add_argument("--ouput_crs", help="CRS for output. Default: 4326 ", type=int, default=4326)
-    parser.add_argument("--ouput_cellsize", help="out put image cellsize. Default: 250 ", type=int, default=250)
+    parser.add_argument("--ouput_cellsize", help="out put image cellsize. Default: 250 ", type=float, default=250)
 
     args = parser.parse_args()
     satellite = args.satellite
